Patching
This is adding go.mod and go.sum files so that it works with Go modules.
It is also fixes #4 and add a length check to avoid it in the future.
It also replaces the old sec51 URL with the http URL from ClamAV since
the other one is down.